add raycaster and player struct initializers

+++ libft/srcs/gfx/ft_argb_lerp.c
@@ -6,18 +6,22 @@
 /*   By: eniini <eniini@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/28 22:04:11 by eniini            #+#    #+#             */
-/*   Updated: 2021/05/28 22:04:20 by eniini           ###   ########.fr       */
+/*   Updated: 2021/06/01 10:18:31 by eniini           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "wolf.h"
+#include "ft_gfx.h"
 
+/*
+*	Interpolates RGB and transparency channels individually between given two
+*	unsigned int color values in 0xAARRGGBB format.
+*/
 uint32_t	ft_argb_lerp(uint32_t c1, uint32_t c2, double p)
 {
-	int a;
-	int r;
-	int g;
-	int b;
+	uint8_t	a;
+	uint8_t	r;
+	uint8_t	g;
+	uint8_t	b;
 
 	if (c1 == c2)
 		return (c1);
